pyproject.toml
[tool.poetry]
name = "guessgame"
version = "0.1.0"
description = ""
authors = ["André Carneiro <aaa@aaa.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.7"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
